Implemented and tested the requirements for the Exercises (4.15, 4.16, 4.17, 4.18, 4.19, 4.20, 4.21, 4.22, 4.23).

+++ controllers/blogs.js
@@ -4,8 +4,8 @@ const Blog = require("../models/blog")
 
 blogsRouter.get("/", async (request, response) => {
     // PART 5 - EXERCISES FOR ASYNC/AWAIT
-    const blogs = await Blog.find()
-    response.json(blogs)
+    const blogs = await Blog.find({}).populate("user", { username: 1, name: 1})
+    response.json(blogs.map(blog => blog.toJSON()))
 
     /*
 
@@ -21,6 +21,8 @@ blogsRouter.get("/", async (request, response) => {
 blogsRouter.post("/", async (request, response) => {
     const body = request.body
 
+    const user = request.user
+
     if (body.title === undefined || body.url === undefined) {
         response.status(400).end()
     }
@@ -29,6 +31,7 @@ blogsRouter.post("/", async (request, response) => {
         title: body.title,
         author: body.author,
         url: body.url,
+        user: user._id
     })
 
     if (body.likes !== undefined){
@@ -41,6 +44,8 @@ blogsRouter.post("/", async (request, response) => {
     // PART 5 EXERCISES FOR ASYNC/AWAIT
     try{
         const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
         response.status(201).json(savedBlog)
     } catch(exception) {
         response.status(400).end()
@@ -58,8 +63,24 @@ blogsRouter.post("/", async (request, response) => {
 
 // EXERCISE (4.13)
 blogsRouter.delete("/:id", async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+
+    const user = request.user
+    const deleteBlog = await Blog.findById(request.params.id)
+
+    if (deleteBlog.user._id.toString() === user._id.toString()) {
+        try{
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } catch (e) {
+            response.status(401).json({
+                error: e.message
+            })
+        }
+    } else {
+        response.status(401).json({
+            error: "Invalid operation / Unauthorized"
+        }).end()
+    }
 })
 
 // EXERCISE (4.14)
